Extracting interface for provider
