+++ app/src/main/java/com/donnfelker/android/bootstrap/BootstrapServiceProvider.java
@@ -1,47 +1,16 @@
-
 package com.donnfelker.android.bootstrap;
 
 import android.accounts.AccountsException;
 import android.app.Activity;
 
-import com.donnfelker.android.bootstrap.authenticator.ApiKeyProvider;
 import com.donnfelker.android.bootstrap.core.BootstrapService;
-import com.donnfelker.android.bootstrap.core.UserAgentProvider;
 
 import java.io.IOException;
 
-import javax.inject.Inject;
-
-import retrofit.RestAdapter;
-
 /**
- * Provider for a {@link com.donnfelker.android.bootstrap.core.BootstrapService} instance
+ * Created by donnfelker on 5/4/15.
  */
-public class BootstrapServiceProvider {
-
-    private RestAdapter restAdapter;
-    private ApiKeyProvider keyProvider;
-
-    public BootstrapServiceProvider(RestAdapter restAdapter, ApiKeyProvider keyProvider) {
-        this.restAdapter = restAdapter;
-        this.keyProvider = keyProvider;
-    }
-
-    /**
-     * Get service for configured key provider
-     * <p/>
-     * This method gets an auth key and so it blocks and shouldn't be called on the main thread.
-     *
-     * @return bootstrap service
-     * @throws IOException
-     * @throws AccountsException
-     */
-    public BootstrapService getService(final Activity activity)
-            throws IOException, AccountsException {
-        // The call to keyProvider.getAuthKey(...) is what initiates the login screen. Call that now.
-        keyProvider.getAuthKey(activity);
-
-        // TODO: See how that affects the bootstrap service.
-        return new BootstrapService(restAdapter);
-    }
+public interface BootstrapServiceProvider {
+    BootstrapService getService(Activity activity)
+            throws IOException, AccountsException;
 }
